+++ R/database_schema.r
@@ -28,11 +28,11 @@
 #'   by \code{\link{synthesise}} or \code{\link{relation_schema}}.
 #' @param references a list of references, each
 #'  represented by a list containing four character elements. In order, the
-#'  elements are a scalar giving the name of the child schema, a vector giving
-#'  the child attribute names, a scalar giving the name of the parent schema,
-#'  and a vector giving the parent attribute names. The vectors must be of the
-#'  same length and contain names for attributes present in their respective
-#'  schemas, and the parent attributes must form a key, in order.
+#'  elements are a scalar giving the name of the child (referrer) schema, a
+#'  vector giving the child attribute names, a scalar giving the name of the
+#'  parent (referee) schema, and a vector giving the parent attribute names. The
+#'  vectors must be of the same length and contain names for attributes present
+#'  in their respective schemas, and the parent attributes must form a key.
 #'
 #' @return A \code{database_schema} object, containing \code{relation_schemas}
 #'   with \code{references} stored in an attribute of the same name.
@@ -197,7 +197,7 @@
 database_schema <- function(relation_schemas, references) {
   if (!inherits(relation_schemas, "relation_schema"))
     stop("relations must be a relation_schema")
-  check_valid_reference(references, relation_schemas, "relation schema")
+  references <- check_valid_reference(references, relation_schemas, "relation schema")
 
   structure(
     relation_schemas,